teams refactor: extracted CreateTeamModal
